requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.6
django-countries==7.6.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
phonenumberslite==8.13.54
pillow==11.1.0
propcache==0.2.1
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
twilio==9.4.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
